adjust layout
